feat(hn_collect) Run at the start of each minute.
Add function to sleep until the next minute starts.
The reason for this is to allow for for the rails runner to be stopped
and started with a known impact, i.e. do whatever needs to be done in a
minute and there is no lose of data. Use the date command to know how
long it is until the next parse of hacker news will take place.